Light refactor, more words, safe R stack
In forth, the return stack is used both for its classic function and as
an auxiliary data stack. This is obviously super dangerous. So, after
thinking about it, I'm providing and auxiliary data stack in the form of
the R stack, which is not used for return address at all and is cleared
with each function. So it's a safe replacement.